chore: reorganisation of ProtectionListener.java
