+++ src/main/java/fr/openmc/core/features/city/commands/AdminCityCommands.java
@@ -1,11 +1,7 @@
 package fr.openmc.core.features.city.commands;
 
 import fr.openmc.api.cooldown.DynamicCooldownManager;
-import fr.openmc.core.features.city.CPermission;
-import fr.openmc.core.features.city.City;
-import fr.openmc.core.features.city.CityManager;
-import fr.openmc.core.features.city.CityMessages;
-import fr.openmc.core.features.city.listeners.ProtectionListener;
+import fr.openmc.core.features.city.*;
 import fr.openmc.core.features.city.mascots.Mascot;
 import fr.openmc.core.features.city.mascots.MascotUtils;
 import fr.openmc.core.features.city.mascots.MascotsManager;
@@ -228,19 +224,19 @@ public class AdminCityCommands {
     @CommandPermission("omc.admins.commands.admincity.claim.bypass")
     public void bypass(Player player) {
         UUID uuid = player.getUniqueId();
-        Boolean canBypass = ProtectionListener.playerCanBypass.get(uuid);
+        Boolean canBypass = ProtectionsManager.playerCanBypass.get(uuid);
 
         if (canBypass == null) {
-            ProtectionListener.playerCanBypass.put(uuid, true);
+            ProtectionsManager.playerCanBypass.put(uuid, true);
             MessagesManager.sendMessage(player, Component.text("Vous pouvez bypass les claims"), Prefix.STAFF, MessageType.SUCCESS, false);
             return;
         }
 
-        if (canBypass == true) {
-            ProtectionListener.playerCanBypass.replace(uuid, false);
+        if (canBypass) {
+            ProtectionsManager.playerCanBypass.replace(uuid, false);
             MessagesManager.sendMessage(player, Component.text("Vous avez désactivé le bypass des claims"), Prefix.STAFF, MessageType.SUCCESS, false);
         } else {
-            ProtectionListener.playerCanBypass.replace(uuid, true);
+            ProtectionsManager.playerCanBypass.replace(uuid, true);
             MessagesManager.sendMessage(player, Component.text("Vous avez activé le bypass des claims"), Prefix.STAFF, MessageType.SUCCESS, false);
 
         }
